requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
pydantic>=2.5.0
